Load, reload and display Locations on the map
Create an ApiService that manages the communication with the API backend.
All other services that need to request the BioPocket API should include
this ApiService.

Create an EnvService that abstract the use of the `ENV` variable. An
EnvInterface is also created, and must be updated each time the `ENV`
variable needs to be updated.
You should include the EnvService wherever you need to access the `ENV`
variable. This helps a lot when creating test as the EnvService can be
easily mocked, which is not the case of the `ENV` variable.

Beware that the `environment.sample.ts` file has changed !
You should update all your `environment.*.ts` files to match this new
sample.

Install `ng-request-builder` to easily make Http requests.
Note that this package will probably be removed when Angular will be
updated to its version 5.*

All services and classes that only export one thing now uses the `default`
keyword, so that we can import them without using destructured
assignments.

The `copy.config.js` file is used to copy assets, stylesheets or scripts
from node modules into the `www` directory, so that they can be referenced
and used at runtime.
See `DEVELOPMENT.md` for more information.

The `Print` utility should be used instead of `console` when you need some
permanent-ish logs, as it does only print logs when allowed to do so.

Story: TG-43