Add support and display for a Map page (#1)
The Leaflet library could be used directly in the project, but doing so
created some strange behavior while navigating through the pages. Replace
the manual use of the lib by the @asymmetrik/ngx-leaflet directive to
solve the problem.

One task asked to remove the list and home page and replace them by the
new map page. Delete only the list page to keep a minimal navigation in
the app. The default home page could be actually deleted when some other
pages will be added. The default page has been changed to the MapPage,
though.

Extending the `config/copy.config.js` script allows us to copy files from
node_module right onto the `build` or `www` folder, for them to be
accessible at runtime. Added the `leaflet.css` on the project this way.

Story: TG-76