Upgrade to Angular 5 (and all other dependencies) (#4)
* Upgrade to Angular 5 (and all other dependencies)

A dependency incompatibility currently exists between the latest
versions of `@ionic/app-scripts` and `postcss`, which is fixed by adding
an additional fixed-version dependency on `postcss` for now. A note has
been added to the development guide to remove this dependency once the
Ionic team has fixed the issue.

For some reason, it appears that the browser tab title is no longer
updated after this upgrade. The e2e test has been updated to take this
into account. It should not be a problem since this is visible only in
development.