Avoid e2e random failure from out-of-bounds marker
The end-to-end test appears to fail randomly. This is probably due to
the screen ratio not being the same as the ratio of the bounding box
within which random locations are generated. The implementation fetches
only locations that are within the screen's bounding box, but the
locations are generated randomly within Onex's bounding box. Some
locations may therefore be outside the screen bounds and not fetched
from the API.

To attempt to fix this issue, the window size has been changed to have
the same width/height ratio as the bounding box within which the
locations are generated, and a padding of 10% has been applied. The
markers should be well within the screen.