Show a card when a location's marker is clicked (#5)
Show a card when a location's marker is clicked

The card is currently rendered using Ionic's Popover service.
Ultimately, this should be switched to a custom implementation ; the
current behavior being not ideal for two main reasons:

* The first is that the animation is relative to the platform and, except
the ios animation, the other are bad for the UX. We should try and
implement our own slide-in up animation.
* The second reason being that when clicking on a location's marker, the map
should move to put this marker at the center of the viewport. But since
the height of the card is needed for the calculation, it's only possible
to move the map *after* the card is displayed, resulting in a delay in the
map animation that is also bad UX.

The card currently display only very basic information without any real
layout. This should be updated when the layout integration starts.

Sidenote: I've added a new section on the `DEVELOPMENT.md` file that list
links to the documentation of libraries used in the project. This should
be updated whenever we add a new major library.

Tests: I've replaced all `callCount` calls with the correct equivalent
from the Chai Sinon Assertion Library (that is : `calledOnce`,
`calledTwice` and so on ; this is for better consistancy.

Story: TG-51
Tasks: TG-99, TG-100