End-to-end test that displays a location's information (#6)
In addition to the test itself, this commit adds utilities to
automatically spawn a backend and mobile application when the end-to-end
tests are run, since the end-to-end tests must execute against the
running application. These utilities can be used locally (see the
documentation) and are used on the Travis continuous integration
environment.

In order to always run the end-to-end tests against a compatible version
of the backend, the backend repository has been added as a submodule of
this repository in the `backend` directory. This will need to be kept
up-to-date so that the end-to-end tests are run on the correct backend
version as new commits are made.

Task: TG-101