Replace legacy Http provider with HttpClient (#7)
Automated tests have had to be updated to the new way of mocking
requests, i.e. using `HttpClientTestingModule` and
`HttpTestingController`.

Formerly, `ApiService` was mostly a utility provider used by other
providers. Now, it's a provider like the others. The generic
functionality it managed (prepending the backend URL to request paths)
is now managed in the new `ApiInterceptorProvider`. Authentication
through the `Authorization` header will later be handled there as well.

Note that unit tests have also been made "more unitary":

* Unit tests for providers should mock HTTP responses, as HTTP calls are
  made in their implementation.
* Unit tests for components should NOT mock HTTP responses, as the calls
  are made in the services. The services themselves should be mocked
  instead. There should be no reason to mock HTTP responses in most (if
  not all) component tests.

The `ng-request-builder` dependency has been removed as it has been made
obsolete by the simpler interface of the new `HttpClient` provider.

An unrelated modification is that the `Geometry` submodel of the
`Location` model was modified to NOT reverse the longitude/latitude
coordinates provided by the API. Since `Geometry` has the same structure
as a GeoJSON point, it's counter-intuitive to not order the coordinates
in the same way. The reversal is done when creating the Leaflet markers
instead, since they use the latitude/longitude order.

Task: TG-107