List available actions with integrated mockups
Add a new component to display each action, named `bip-action-card`.
It will be reused when the dashboard and the suggested actions will be
implemented.
It will also be updated when action microdata (like cost, time, etc) will
be implemented.

Add basis navigation for each Action. Right now, a simple blank page
displaying the name of the Action is available. The next PR will implement
the complete page (based on the currently available Action's data).

Update test to add some for the new component and update existing ones.
Also update the e2e, although on a very tiny scale.

STORY: TG-52