Adds an Action page
Adds a new directive to manage broken image link and replace them with a
default image, stored locally.
Also added a new typescript alias to the new `directives` folder.

Integrates the mockup for the Action page.

Adds a new page for the Theme, but it's currently blank. This will be
implemented in the following PR.

Adds some native plugins to the app, that are not necessarily used now
(like the Toast one) but will be in a not so distant future.

Implements some features regarding the Status Bar, that will could be
enhanced in future PR.

Updates unit and e2e test to include the new Action page.
Also makes small updates on previously existing tests.