Integrate theme description page mockup.
Update CSS and markup for the Theme description page so that it matches
the mockups.

Add a basic error handling in the map, specifically when the
geolocation takes too much time (either because the app does not have
sufficient permissions or the GPS is not easily available).

This PR also includes changes related to testing the app on an iOS device.
This resulted in adding configuration for the ios platform in the project
files, update the app name and reference in the `config.xml` because,
following a mistake of mine, the app on Apple Store and the one on the
Google Play does not have the same bundle ID :
* On App Store, it's `ch.heig-vd.biopocket`
* On Google Play, it's `biopocket.mediamaps.ch`

Separate app icons have been added to the project (both Google Play and
App Store have different requirements concerning the app icon).

Fix some bugs:
* When building the app for production with `ionic cordova build
  <platform> --prod`, the process failed due to some exports made with
  `export default`. As a result, all `export default` have been removed
  and replaced by regular exports.
* The `image-handler` directive had a bug that replaced every pre-existing
  CSS classes on the host component by the one class the directive added.

The e2e test have been updated to include two new steps:
* Clicking on the action theme and going to the corresponding Theme page
* Going back from the Theme page to the previous Action page.

Story: TG-56