Add a registration module
Add a new module to the app that contains all features to allow the app's
user to register in order to receive news on the project later on.

Those new features have been grouped on a new module to allow easy
removing of the features when not needed (once user can create an account,
this register feature will become useless).

The module contains several new components:
* A `register-form` component that manages the logic for user registration
* An `unregister-form` component taht manages the logic for user
  unregistration
* A `RegistrationPage` to display the `register-form`
* An `UnregistrationPage` to display the `unregister-form`
* A `RegistrationTabPage` to display both preceeding pages with tabs
* A `RegistrationService` that handle the backend communication
* A `RegistrationModal` that displays the `register-form` when a user
  tries to participate to an action

A directive have also been created to manage the state of the email form
input in both the forms.

Most of those new components have their respective tests.

This PR also remove the fake data used on the main menu or as placeholder
images on the action and theme, and replaces theme with more adequate data
and images.

Several new Cordova plugins have been added to manage certains features
related to those registration features.

STORY: TG-152