Home Page content and markdown parser (#16)
Change the home page content so that it now displays information about the
BioPocket project, along with the YouTube video produced by Melanie.
This new Home Page is now the default page of the app.

To include the YouTube video, a new Ionic Native plugin has been added to the project and its usage documented.

As a result, the main e2e test has been updated to match this new state by
adding some step at the start of the test.

A second e2e test has been added to the e2e test suite. This has been done
in order not to change too much the main scenario ; the registration
process being only temporary, until the user account are implemented.

Create a new method expectTitle() in the AbstractPageObject class,
as discussed here:
https://github.com/MediaComem/biopocket-mobile/pull/14#discussion_r236686807

Add two new npm scripts: `ipr` (Ionic Production Release) and `apr`
(Android Production Release), each of them launching ionic command to
build the app for production for a platform (`ipr` for iOS and `apr` for
Android).
Note that `ipr` include the build flag for cordova not being compatible
with XCode 10+.

Also adds two dummy scripts, `zipalign` and `apksigner` that are just
reminders of the command to execute when signing an apk for Android.
Ideally, there should be a script that executes all the command in the
Android building process, based on a configuration containing
paths and passwords.

This PR also fix some bugs on iOS, namely the one where the status bar was
overlapping the app content, and the one where the app was not asking for
geolocation permission.

The folder `classes` is remove due to being a big redundancy with the
`models` folder (that stays).
All the references to the files of this folder have been updated.
The `@classes` alias has been removed from the tsconfig and webpack config.

Also contains some various fixes for the tests and miscellaneous changes
to the code.

Witht this PR, the app version should be bumped to v1.0.0!